introduce engine-solution module
